refactor: move the notification related logic to ChangelogTrigger.js
